added comments; float bug fix for linewidth, zorder, and (for some reason) linestyle

+++ qpp.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python
-##  head -n 17 qpp.py
+##  head -n 19 qpp.py
 ##  python qpp.py config.in output_plot.pdf
 ##  By: Charlie Payne, 2018-2019
 ## DESCRIPTION
@@ -11,6 +11,8 @@
 ##  -- it would be nice to somehow have default values for the 'kscales' and alike, for now rely on default config.in
 ##  -- somehow generalize kopts such that the user can use any general python function on the data
 ##  -- maybe automate a line border
+##  -- add in the 'alpha' option
+##  -- add in error bar plotting
 ##  -- somehow make a sub-figure setup... :|
 ## PARAMETERS
 ##  1) confile = sys.argv[1]     # the plotting config file, see 'config.in' for the default structure
@@ -119,13 +121,13 @@ kmin = 0  # the lower-bound index for the data points per curve in the concatena
 for k in range(knum):
   kmax = lens[k] + kmin  # the upper-bound index " " " " " " " " " " "
   if axesopt == 'linear':
-    plot(data[kmin:kmax,0], data[kmin:kmax,1], klines[str(k)], color=kcolors[str(k)], label=klabs[str(k)], linewidth=kwidths[str(k)], zorder=int(korders[str(k)]))
+    plot(data[kmin:kmax,0], data[kmin:kmax,1], klines[str(k)], color=kcolors[str(k)], label=klabs[str(k)], linewidth=float(kwidths[str(k)]), zorder=float(korders[str(k)]))
   elif axesopt == 'semilogy':
-    semilogy(data[kmin:kmax,0], data[kmin:kmax,1], klines[str(k)], color=kcolors[str(k)], label=klabs[str(k)], linewidth=kwidths[str(k)], zorder=int(korders[str(k)]))
+    semilogy(data[kmin:kmax,0], data[kmin:kmax,1], klines[str(k)], color=kcolors[str(k)], label=klabs[str(k)], linewidth=float(kwidths[str(k)]), zorder=float(korders[str(k)]))
   elif axesopt == 'semilogx':
-    semilogx(data[kmin:kmax,0], data[kmin:kmax,1], klines[str(k)], color=kcolors[str(k)], label=klabs[str(k)], linewidth=kwidths[str(k)], zorder=int(korders[str(k)]))
+    semilogx(data[kmin:kmax,0], data[kmin:kmax,1], klines[str(k)], color=kcolors[str(k)], label=klabs[str(k)], linewidth=float(kwidths[str(k)]), zorder=float(korders[str(k)]))
   elif axesopt == 'loglog':
-    loglog(data[kmin:kmax,0], data[kmin:kmax,1], klines[str(k)], color=kcolors[str(k)], label=klabs[str(k)], linewidth=kwidths[str(k)], zorder=int(korders[str(k)]))
+    loglog(data[kmin:kmax,0], data[kmin:kmax,1], klines[str(k)], color=kcolors[str(k)], label=klabs[str(k)], linewidth=float(kwidths[str(k)]), zorder=float(korders[str(k)]))
   else:
     print 'ERROR 0798: figparms[axes] not recognized!'
     print 'figparms[axes] =',axesopt
